requirements.txt
fastapi
uvicorn
pydantic
asyncpg
python-dotenv
upstash-redis
pytest-asyncio 
httpx
redis 
requests 
pika
pybreaker
sendgrid
python-http-client
psycopg2-binary
sqlalchemy
python-multipart
alembic>
python-jose[cryptography]
passlib[bcrypt]
gunicorn
# firebase-admin
tenacity